(chore) move list-related hook to components/List/hooks
