+++ frontend/src/components/list/hooks/useList.tsx
@@ -1,12 +1,12 @@
+import { listState } from "components/list/state/listAtoms";
 import { numTermsToReviewState } from "components/review/state/reviewSelectors";
 import { List } from "gql/codegen-output";
 import { useQueryListsById } from "gql/hooks/list-query";
 import { useRouteProps } from "hooks/routerHooks";
 import { useCallback, useEffect, useState } from "react";
 import { useRecoilValue, useResetRecoilState, useSetRecoilState } from "recoil";
-import { listState } from "state/atoms/listAtoms";
 import { TruncatedTerm } from "types/list.types";
-import ListTerm from "../../components/list/sub/ListTerm";
+import ListTerm from "../sub/ListTerm";
 import { useListFilter } from "./useListFilter";
 import { useListPrepareReview } from "./useListPrepareReview";
 import { useListUpdate } from "./useListUpdate";
