+++ frontend/src/components/list/hooks/useListPrepareReview.ts
@@ -1,3 +1,4 @@
+import { selectingTermsToReviewState } from "components/list/state/listAtoms";
 import { termsToReviewState } from "components/review/state/reviewAtoms";
 import { List } from "gql/codegen-output";
 import { Maybe } from "graphql/jsutils/Maybe";
@@ -5,7 +6,6 @@ import { filterTermsBySaturation } from "helpers/filterTermsBySaturation";
 import { suggestTermsForReview } from "helpers/srs/saturation";
 import { useCallback, useMemo } from "react";
 import { useRecoilState, useResetRecoilState, useSetRecoilState } from "recoil";
-import { selectingTermsToReviewState } from "state/atoms/listAtoms";
 import { FilterInterface } from "types/list.types";
 
 type Args = {
